Fixed Pic Component and Inputs
Added state management for the inputs for the export functionality. The FormComponent component will now store the data from the MatInputs in its state so that the export component will be able to grab the data and generate a new component without the text boxes or Change Gradient button and such.